# optimize index computation