--- a/paddle/fluid/operators/roll_op.cu
+++ b/paddle/fluid/operators/roll_op.cu
@@ -36,12 +36,12 @@ __global__ void RollCudaKernel(const T* input, T* output, int64_t N,
   }
 
   int64_t output_idx = idx;
-  int64_t dim_idx, dim_idx_shift;
+  int64_t new_dim_idx = 0;
 
-#pragma unroll Rank
+#pragma unroll
   for (size_t i = 0; i < Rank; i++) {
-    dim_idx = (idx / strides[i]) % sizes[i] + shifts[i];
-    if (dim_idx >= sizes[i]) {
+    new_dim_idx = (idx / strides[i]) % sizes[i] + shifts[i];
+    if (new_dim_idx >= sizes[i]) {
       output_idx += (shifts[i] - sizes[i]) * strides[i];
     } else {
       output_idx += shifts[i] * strides[i];