<!-- Demo: https://github.com/PaddlePaddle/Paddle/pull/24810 -->
### PR types
<!-- One of [ New features | Bug fixes | Function optimization | Performance optimization | Breaking changes | Others ] -->
Performance optimization

### PR changes
<!-- One of [ OPs | APIs | Docs | Others ] -->
OPs

### Describe
<!-- Describe what this PR does -->
optimize the index computation

**Paddle vs Pytorch**
| --- | Paddle-Old | Pytorch | Paddle |
|---|---|---|---|
| shape=(1000,2000)<br> axis=(1)<br>shift=(5) | avg ~ 25.5 us | avg ~ 24us | avg ~ 23.75us |
| shape=(1000,2000)<br> axis=(0)<br>shift=(5) | avg ~ 24.9us | avg ~ 24us | avg ~ 22.8us |
| shape=(1000,2000)<br> axis=(0,1)<br>shift=(5,5) | avg ~ 37.7 us | avg ~ 48us | avg ~ 31.4us |